Revert "Update .htaccess"
This reverts commit b7fe7ce7b875391265ec0c3f65c2fcfa57cae7aa.
